Move `FrameBox` to its own file
