+++ src/pdu_loop/pdu_tx.rs
@@ -1,5 +1,6 @@
 use super::{
-    frame_element::{sendable_frame::SendableFrame, FrameBox, FrameElement},
+    frame_box::FrameBox,
+    frame_element::{sendable_frame::SendableFrame, FrameElement},
     storage::PduStorageRef,
 };
 use core::{ptr::NonNull, task::Waker};
